Opcode ADD, INC, LDH, and small fixes
Implemented opcodes for 0x80--0x87 (ADD), 04--34 & 0C--3C (INC), and
opcode 0xE0 (LHD n, A). This is part of implementing sufficient opcodes
for the bootloader without implementing all of them at once. Where
possible opcodes are generalized (such as ADD and INC being implemented
block-by-block rather than individually such as 0xE0).

Many small lint fixes are included in this commit such as semicolons
etc. Linting is in compliance with esversion 6 and JSHint is being used
for this purpose.